Add Authentication
I was tried to make a extension class of NodeRtmpSession.. but it failed.
So I modify it to authenticate based on RTMP handshake packet.
It just pass the data into Jupiter's Stream ID Manager and waits for response.
Implementing with callback function help me learn about callback function. thx for Google.